requirements.txt
openai>=1.30.0
chromadb>=0.5.0
streamlit>=1.35.0
tiktoken>=0.7.0
numpy>=1.26.0
scikit-learn>=1.5.0
python-dotenv>=1.0.0
